Funções de menu e Jogo forca modularizados
